% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mul.sdwd.R
\name{mul.sdwd}
\alias{mul.sdwd}
\title{Multiway Sparse Distance Weighted Discrimination}
\usage{
mul.sdwd(X, y, R = 1, lambda1 = 0, lambda2 = 1, convThresh = 1e-05, nmax = 500)
}
\arguments{
\item{X}{A multiway array with dimensions \eqn{N \times P_1 \times ... \times P_K}. The first dimension (N) give the cases (e.g.subjects) to be classified.}

\item{y}{A vector of length N with class label (-1 or 1) for each case.}

\item{R}{Assumed rank of the coefficient array.}

\item{lambda1}{The L1 tuning parameter lambda1 that enforce sparsity. lambda1 and lambda2 can be determined by the function \code{\link{lambda.mul.sdwd}}.}

\item{lambda2}{The L2 tuning paprameter lambda2.}

\item{convThresh}{The algorithm stops when the distance between B_new and B_old is less than convThresh. Default is 1e-7.}

\item{nmax}{Restrics how many iterations are allowed. Default is 500.}
}
\value{
A list of components

\code{beta} Vector of coefficients with length \eqn{P_1 \times ... \times P_K}.

\code{int} Intercept.

\code{U} A list of K matrices. Each matrix (\eqn{P_k \times R}) corresponds to the estimated weights for \eqn{k^th} dimension.

If R=1, U is a list of K vectors.
}
\description{
Optimize sparse DWD objective for a multiway dataset with any dimension. L1 and L2 parameters are imposed to enforce sparsity in the model.
}
\examples{
## Load gene expression time course data (?IFNB_Data for more info)
data(IFNB_Data)
Class=2*(Class-1)-1 #redefine classes to -1 or 1

## Run Multiway SDWD
res.msdwd1 <- mul.sdwd(DataArray,y=Class,R=1,lambda1=0,lambda2=1,convThresh = 1e-5, nmax=500)

##Compute projection onto the classification direction for each individual:
scores <- c()
for(i in 1:length(Class)) scores[i] = sum(as.vector(DataArray[i,,])*res.msdwd1$beta)+res.msdwd1$int
Colors=rep(1,length(Class))
Colors[Class==1]=2
plot(scores, col=Colors)
}
